unify all uses of the retry client
