+++ nexus/internal/uploader/uploader_default.go
@@ -1,50 +1,16 @@
 package uploader
 
 import (
-	"log/slog"
 	"net/http"
 	"os"
 	"strings"
 	"time"
 
+	"github.com/wandb/wandb/nexus/internal/clients"
+
 	"github.com/hashicorp/go-retryablehttp"
 	"github.com/wandb/wandb/nexus/pkg/observability"
 )
-
-type RetryClientOption func(rc *retryablehttp.Client)
-
-func WithRetryClientLogger(logger *observability.NexusLogger) RetryClientOption {
-	return func(rc *retryablehttp.Client) {
-		rc.Logger = slog.NewLogLogger(logger.Logger.Handler(), slog.LevelDebug)
-	}
-}
-
-func WithRetryClientRetryMax(retryMax int) RetryClientOption {
-	return func(rc *retryablehttp.Client) {
-		rc.RetryMax = retryMax
-	}
-}
-
-func WithRetryClientRetryWaitMin(retryWaitMin time.Duration) RetryClientOption {
-	return func(rc *retryablehttp.Client) {
-		rc.RetryWaitMin = retryWaitMin
-	}
-}
-
-func WithRetryClientRetryWaitMax(retryWaitMax time.Duration) RetryClientOption {
-	return func(rc *retryablehttp.Client) {
-		rc.RetryWaitMax = retryWaitMax
-	}
-}
-
-func NewRetryClient(opts ...RetryClientOption) *retryablehttp.Client {
-	retryClient := retryablehttp.NewClient()
-
-	for _, opt := range opts {
-		opt(retryClient)
-	}
-	return retryClient
-}
 
 // DefaultUploader uploads files to the server
 type DefaultUploader struct {
@@ -57,15 +23,14 @@ type DefaultUploader struct {
 
 // NewDefaultUploader creates a new uploader
 func NewDefaultUploader(logger *observability.NexusLogger) *DefaultUploader {
-	retryClient := NewRetryClient(
-		WithRetryClientLogger(logger),
-		WithRetryClientRetryMax(10),                 // todo: make this configurable
-		WithRetryClientRetryWaitMin(1*time.Second),  // todo: make this configurable
-		WithRetryClientRetryWaitMax(60*time.Second), // todo: make this configurable
+	retryClient := clients.NewRetryClient(
+		clients.WithRetryClientLogger(logger),
+		clients.WithRetryClientRetryMax(10),                 // todo: make this configurable
+		clients.WithRetryClientRetryWaitMin(1*time.Second),  // todo: make this configurable
+		clients.WithRetryClientRetryWaitMax(60*time.Second), // todo: make this configurable
 	)
 
 	// todo: add ResponseHandlerFunc to retryClient to handle non-200 responses
-
 	uploader := &DefaultUploader{
 		client: retryClient,
 		logger: logger,
